First major commit
This commit has working and linked GUI components for employees and
customers as we are still unsure of how we are going to implement the
expenses and inventory relationships. For the moment there is NO
database connection setup, but entry into the fields will be populated
to the appropreate tables and I have inserted dummy data where needed.