Added some inventory things and bug fixes
Added an inventory viewer based on Preston's schemea. It looks almost
identical to the customer and employee entries and works largely the
same way. I dont know if there should be anything different for these.

Also implement a few bug fixes in BaseGUI so that the same model is not
re-instantiated every time one of the tables is updated. In the new way
there is a new model for each dialog that requires entry. I do not know
if this is better, but it seems like a good idea to avoid collisions.